stable-diffusion-wrapper/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.0.1+cpu
torchvision==0.15.2+cpu
diffusers==0.21.0
transformers==4.30.0
accelerate==0.20.0
pillow==10.0.0
safetensors==0.3.1